Write modified buffer to a tmp file to work around Vim issue
Setting `noblock` on the channel doesn't help; since the stdin
is in TTY mode, the Node process seems to not produce any data
from it until it hits a newline. Since Node isn't consuming the
data, if the file is longer than Vim's buffer then Vim will hang
trying to write to the process.