Create a "Simple" runner that starts a new term each time
As the name implies, this is a much simpler interaction. I wasn't
happy with it when I tried it initially, but it works okay. It does
feel a little slower than the daemon, but that could just be because
there's no initial output while the request runs.

The daemon runner just wasn't working the way we want it to, since
long responses would get left in the buffer's history, and Vim doesn't
allow us to edit the buffer of a running job.